requirements.txt
Flask==3.0.3
huggingface_hub>=0.24.0
python-dotenv==1.0.1